// location-finder/src/app/models/location.model.ts
export interface LocationDetails {
  fullAddress: string;
  streetNumber: string;
  streetName: string;
  areaName: string;
  city: string;
  district: string;
  state: string;
  country: string;
  postalCode: string;
  latitude: number;
  longitude: number;
  placeId: string;
  formattedAddress: string;
}